% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FortranDemo-package.R
\docType{package}
\name{FortranDemo-package}
\alias{FortranDemo}
\alias{FortranDemo-package}
\title{FortranDemo: Calculating the sqaure and cubic values}
\description{
This package implements a function for calculating the squared and cubic value of
    an integer value or a vector of integer values.
}
\keyword{internal}
