% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rinterface.r
\name{square.cube}
\alias{square.cube}
\title{Squared and cubic values}
\usage{
square.cube(x)
}
\arguments{
\item{x}{the positive integer}
}
\value{
A list of 2, the first element is the squared value, the second is the cubic value.
}
\description{
This function calculates the squared and cubic values of a positive integer value.
}
\details{
This function requires a single argument, which should be a positive integer value.
}
\examples{
square.cube(3)
square.cube(11)
\dontrun{square.cube(123)}

}
\references{
Reference 1...

Reference 2...
}
